extract `reduce_pipeline`
